+++ rsapi/http.go
@@ -23,6 +23,71 @@ type FileUpload struct {
 	Reader   io.Reader // Backing reader
 }
 
+// SourceUpload represents a payload for /api/right_scripts/:id/update_source
+// It's needed as this one API call has unique properties -- it's content type is
+// text/plain only, and body of the HTTP call is the script body rather than a properly
+// formatted form request :(
+type SourceUpload struct {
+	Reader io.Reader
+}
+
+// This will take a hash (or multi-level hash) of APIParams and extract
+// out all the multipart file uploads and put them in an array. The recursion
+// is needed as CM API 1.5 and SS both use this, but SS has all its params
+// on the top level while CM API 1.5 has the resource name on the top level
+// with the params being one level down the tree.
+func extractUploads(payload *APIParams) (uploads []*FileUpload) {
+	for k, v := range *payload {
+		if mpart, ok := v.(*FileUpload); ok {
+			uploads = append(uploads, mpart)
+			delete(*payload, k)
+		} else if child_params, ok := v.(APIParams); ok {
+			child_uploads := extractUploads(&child_params)
+			uploads = append(uploads, child_uploads...)
+		}
+	}
+	return uploads
+}
+
+func extractSourceUpload(payload *APIParams) *SourceUpload {
+	var sourceUpload *SourceUpload
+	for k, v := range *payload {
+		if mpart, ok := v.(*SourceUpload); ok {
+			sourceUpload = mpart
+			delete(*payload, k)
+		}
+	}
+	return sourceUpload
+}
+
+// Handle payload params. Each payload param gets its own multipart
+// form section with the section name being the variable name and
+// section contents being the variable contents. Handle recursion as well.
+func writeMultipartParams(w *multipart.Writer, payload APIParams, prefix string) error {
+	for k, v := range payload {
+		fieldName := k
+		if prefix != "" {
+			fieldName = fmt.Sprintf("%s[%s]", prefix, k)
+		}
+		// Add more types as needed. These two cover both CM15 and SS.
+		switch v.(type) {
+		case string:
+			err := w.WriteField(fieldName, v.(string))
+			if err != nil {
+				return err
+			}
+		case APIParams:
+			err := writeMultipartParams(w, v.(APIParams), fieldName)
+			if err != nil {
+				return err
+			}
+		default:
+			return fmt.Errorf("Unknown type for multipart form section %s: %#v", fieldName, v)
+		}
+	}
+	return nil
+}
+
 // BuildHTTPRequest creates a http.Request given all its parts.
 // If any member of the Payload field is of type io.Reader then the resulting request has a
 // multipart body where each member of type io.Reader is mapped to a single part and all other
@@ -68,16 +133,13 @@ func (a *API) BuildHTTPRequest(verb, path, version string, params, payload APIPa
 	var jsonBytes []byte
 	var body io.Reader
 	var isMultipart bool
+	var isSourceUpload bool
 	var boundary string
 	if payload != nil {
 		var fields io.Reader
-		var uploads []*FileUpload
-		for k, v := range payload {
-			if mpart, ok := v.(*FileUpload); ok {
-				uploads = append(uploads, mpart)
-				delete(payload, k)
-			}
-		}
+		uploads := extractUploads(&payload)
+		sourceUpload := extractSourceUpload(&payload)
+
 		if len(payload) > 0 {
 			var err error
 			if jsonBytes, err = json.Marshal(payload); err != nil {
@@ -90,13 +152,9 @@ func (a *API) BuildHTTPRequest(verb, path, version string, params, payload APIPa
 			var buffer bytes.Buffer
 			w := multipart.NewWriter(&buffer)
 			if len(payload) > 0 {
-				p, err := w.CreateFormField("payload")
+				err := writeMultipartParams(w, payload, "")
 				if err != nil {
-					return nil, fmt.Errorf("failed to create multipart payload: %s", err.Error())
-				}
-				_, err = io.Copy(p, fields)
-				if err != nil {
-					return nil, fmt.Errorf("failed to copy multipart payload: %s", err.Error())
+					return nil, fmt.Errorf("failed to write multipart params: %s", err.Error())
 				}
 			}
 			for _, u := range uploads {
@@ -114,6 +172,9 @@ func (a *API) BuildHTTPRequest(verb, path, version string, params, payload APIPa
 			}
 			boundary = w.Boundary()
 			body = &buffer
+		} else if sourceUpload != nil {
+			isSourceUpload = true
+			body = sourceUpload.Reader
 		} else {
 			body = fields
 		}
@@ -127,6 +188,8 @@ func (a *API) BuildHTTPRequest(verb, path, version string, params, payload APIPa
 	}
 	if isMultipart {
 		req.Header.Set("Content-Type", fmt.Sprintf("multipart/form-data; boundary=%s", boundary))
+	} else if isSourceUpload {
+		req.Header.Set("Content-Type", "text/plain")
 	} else {
 		req.Header.Set("Content-Type", "application/json")
 	}